requirements.txt
Django==1.10.5
django-material==0.12.5
djangorestframework==3.5.4
django-allauth
djangorestframework-jwt